web/api/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
psycopg2-binary>=2.9
google-auth>=2.29
pyjwt>=2.8
python-multipart>=0.0.9
python-dotenv>=1.0
sqlalchemy>=2.0
